journal-my-life-api/requirements.txt
flask
python-dotenv
jsonify
gunicorn
flask-cors
datetime
requests
